Refactor map geometry fetching to hooks
